OMAP4: intruduce silicon type override when set, this enables basic overclocking capabilites. use with care!
most (if not all) of the omap4xxx SOCs should be able to reach higher frequencies. with smartreflex and tempgov in place this should also be safe on the respective platforms.

in omap4_check_features(void) OMAP4_SILICON_TYPE_PERFORMANCE is enforced, which enables higher MPU (on 446x also IVA) frequencies regardless of the actual silicon type read from die and production IDs